Update Chat.jsx
Im not sure if this code will work now, but what I'm trying to do is broadcast information about the tip being made. For now it is broadcasted locally and I think its wrong. Shouldn't be like that. Tips can be used to tip people that are helpful or funny, but it makes more sense when info about that event is broadcasted to all, otherwise it feels creepy.
I think automatic memo is also better aproach,  less effort for sender, but he can change if he wish.
This scheme could go further with for example Options: defaultTip = 100,  defaultToken = BTS . That way user effort is limited to simple /tip user even is he wish use nonstandard values.


Those parameters can be also overwrited as user is typing:

 /tip user notDefaultValue
/tip user notDefaultValue notDefaultValue
and so on...